Extract confirmation to new controller
